Revert "Feature/other branch reversioning extension"
