+++ pom.xml
@@ -8,7 +8,7 @@
 
     <groupId>com.e-gineering</groupId>
     <artifactId>gitflow-helper-maven-plugin</artifactId>
-    <version>3.1.0-SNAPSHOT</version>
+    <version>3.0.0</version>
 
     <packaging>maven-plugin</packaging>
