+++ src/main/java/com/e_gineering/maven/gitflowhelper/RetargetDeployMojo.java
@@ -1,5 +1,9 @@
 package com.e_gineering.maven.gitflowhelper;
 
+import org.apache.maven.artifact.Artifact;
+import org.apache.maven.artifact.InvalidRepositoryException;
+import org.apache.maven.artifact.versioning.VersionRange;
+import org.apache.maven.model.DeploymentRepository;
 import org.apache.maven.model.DistributionManagement;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.MojoFailureException;
@@ -37,7 +41,16 @@ public class RetargetDeployMojo extends AbstractGitflowBasedRepositoryMojo {
                 String otherBranchesToDeploy = resolveExpression(otherDeployBranchPattern);
 	            if (!"".equals(otherBranchesToDeploy) && gitBranchInfo.getName().matches(otherBranchesToDeploy)) {
                     setTargetSnapshots();
-                    // Only break if we set something!
+
+                    project.setVersion(getAsBranchSnapshotVersion(project.getVersion(), gitBranchInfo.getName()));
+
+                    // Update any attached artifacts.
+                    updateArtifactVersion(project.getArtifact(), gitBranchInfo.getName());
+                    for (Artifact a : project.getAttachedArtifacts()) {
+                        updateArtifactVersion(a, gitBranchInfo.getName());
+                    }
+
+                    getLog().info("Artifact versions updated with build metadata: " + getAsBranchSnapshotVersion("", gitBranchInfo.getName()));
                     break;
                 }
             }
@@ -46,6 +59,35 @@ public class RetargetDeployMojo extends AbstractGitflowBasedRepositoryMojo {
                 break;
             }
         }
+    }
+
+    /**
+     * Updates artifact versions for a given branch name.
+     * @param a artifact to update (may be null)
+     * @param branchName the branch name
+     */
+    private void updateArtifactVersion(Artifact a, String branchName) {
+        if (a != null) {
+            a.setVersion(getAsBranchSnapshotVersion(a.getVersion(), branchName));
+            try {
+                a.setVersionRange(VersionRange.createFromVersion(a.getVersion()));
+            } catch (UnsupportedOperationException uoe) { // Some artifact types don't like this.
+                getLog().debug("Unable to update VersionRange for artifact.");
+            }
+        }
+    }
+
+
+    /**
+     * Given a String version (which may be a final or -SNAPSHOT version) return a
+     * version version string mangled to include a `+normalized-branch-name-SNAPSHOT format version.
+     *
+     * @param version The base version (ie, 1.0.2-SNAPSHOT)
+     * @param branchName to be normalized
+     * @return A mangled version string with the branchname and -SNAPSHOT.
+     */
+    private String getAsBranchSnapshotVersion(final String version, final String branchName) {
+        return version.replace("-SNAPSHOT", "") + otherBranchVersionDelimiter + branchName.replaceAll("[^0-9A-Za-z-.]", "-") + "-SNAPSHOT";
     }
 
     private void setTargetSnapshots() throws MojoExecutionException, MojoFailureException {
