+++ src/main/java/com/e_gineering/maven/gitflowhelper/properties/CircularDefinitionPreventer.java
@@ -19,11 +19,7 @@ package com.e_gineering.maven.gitflowhelper.properties;
  * under the License.
  */
 
-import java.util.HashSet;
-import java.util.Iterator;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Set;
+import java.util.*;
 
 /**
  * Shamelessly cribbed from https://raw.githubusercontent.com/mojohaus/properties-maven-plugin/master/src/main/java/org/codehaus/mojo/properties/CircularDefinitionPreventer.java
