+++ src/test/java/com/e_gineering/maven/gitflowhelper/RetargetDeployIT.java
@@ -123,6 +123,7 @@ public class RetargetDeployIT extends AbstractIntegrationTest {
 
 	@Test
 	public void othersUnsetRepos() throws Exception {
+		// Deploy a hotfix to the test-releases.
 		Verifier verifier = createVerifier("/project-stub", "feature/undeployable", "3.5.0-SNAPSHOT");
 		try {
 			verifier.executeGoal("deploy");
