--- src/test/resources/multi-module-project-stub/child1/src/main/java/HelloWorld.java
@@ -1,8 +1,0 @@
-/**
- * Simple HelloWorld to stuff into a .jar for this project-stub.
- */
-public class HelloWorld {
-	public static void main(String[] args) {
-		System.out.println("Hello, World.");
-	}
-}
