--- src/test/resources/multi-module-project-stub/child1/src/main/resources/project.properties
@@ -1,1 +1,0 @@
-app.buildinfo=${build.branch.string}
