--- src/test/resources/multi-module-project-stub/child2/pom.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-
-	<groupId>com.e-gineering</groupId>
-	<artifactId>gitflow-helper-maven-plugin-multi-module-child2-stub</artifactId>
-<!--	<version>${stub.project.version}</version>-->
-
-	<packaging>jar</packaging>
-
-	<parent>
-		<groupId>com.e-gineering</groupId>
-		<artifactId>gitflow-helper-maven-plugin-multi-module-parent-stub</artifactId>
-		<version>${stub.project.version}</version>
-	</parent>
-
-	<dependencies>
-		<dependency>
-			<groupId>com.e-gineering</groupId>
-			<artifactId>gitflow-helper-maven-plugin-multi-module-child1-stub</artifactId>
-			<version>${stub.project.version}</version>
-		</dependency>
-	</dependencies>
-</project>
