--- src/test/resources/multi-module-project-stub/pom.xml
@@ -1,134 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-
-	<groupId>com.e-gineering</groupId>
-	<artifactId>gitflow-helper-maven-plugin-multi-module-parent-stub</artifactId>
-	<version>${stub.project.version}</version>
-
-	<packaging>pom</packaging>
-
-	<modules>
-		<module>child1</module>
-		<module>child2</module>
-	</modules>
-
-	<properties>
-		<maven.compiler.source>1.8</maven.compiler.source>
-		<maven.compiler.target>1.8</maven.compiler.target>
-		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-	</properties>
-
-	<repositories>
-		<repository>
-			<id>releases</id>
-			<url>file:${it.repository.basedir}/releases</url>
-			<snapshots><enabled>false</enabled></snapshots>
-			<releases><enabled>true</enabled></releases>
-		</repository>
-		<repository>
-			<id>test-releases</id>
-			<url>file:${it.repository.basedir}/test-releases</url>
-			<snapshots><enabled>false</enabled></snapshots>
-			<releases><enabled>true</enabled></releases>
-		</repository>
-		<repository>
-			<id>snapshots</id>
-			<url>file:${it.repository.basedir}/snapshots</url>
-			<snapshots><enabled>true</enabled></snapshots>
-			<releases><enabled>false</enabled></releases>
-		</repository>
-	</repositories>
-
-	<build>
-		<resources>
-			<resource>
-				<directory>src/main/resources</directory>
-				<filtering>true</filtering>
-				<includes>
-					<include>*.properties</include>
-				</includes>
-			</resource>
-		</resources>
-
-		<pluginManagement>
-			<plugins>
-				<plugin>
-					<groupId>org.codehaus.mojo</groupId>
-					<artifactId>build-helper-maven-plugin</artifactId>
-					<version>3.0.0</version>
-					<executions>
-						<execution>
-							<id>attach-text-file</id>
-							<goals>
-								<goal>attach-artifact</goal>
-							</goals>
-							<configuration>
-								<artifacts>
-									<artifact>
-										<file>src/build/resources/my-text-file.txt</file>
-										<type>text-file</type>
-									</artifact>
-								</artifacts>
-							</configuration>
-						</execution>
-					</executions>
-				</plugin>
-			</plugins>
-		</pluginManagement>
-
-		<plugins>
-			<plugin>
-				<groupId>com.e-gineering</groupId>
-				<artifactId>gitflow-helper-maven-plugin</artifactId>
-				<version>${version.gitflow.plugin}</version>
-				<extensions>true</extensions>
-				<configuration>
-					<releaseDeploymentRepository>releases</releaseDeploymentRepository>
-					<stageDeploymentRepository>test-releases</stageDeploymentRepository>
-					<snapshotDeploymentRepository>snapshots</snapshotDeploymentRepository>
-					<masterBranchPropertyFile>src/build/resources/master.build.properties</masterBranchPropertyFile>
-					<supportBranchProperties>
-						<build.branch.type>support</build.branch.type>
-					</supportBranchProperties>
-					<releaseBranchProperties>
-						<build.branch.type>release</build.branch.type>
-					</releaseBranchProperties>
-					<hotfixBranchProperties>
-						<build.branch.type>hotfix</build.branch.type>
-					</hotfixBranchProperties>
-					<developmentBranchProperties>
-						<build.branch.type>develop</build.branch.type>
-					</developmentBranchProperties>
-					<otherBranchProperties>
-						<build.branch.type>other</build.branch.type>
-					</otherBranchProperties>
-					<undefinedBranchProperties>
-						<build.branch.type>undefined</build.branch.type>
-					</undefinedBranchProperties>
-					<otherDeployBranchPattern>(origin/)?feature/poc/.*</otherDeployBranchPattern>
-				</configuration>
-				<executions>
-					<execution>
-						<goals>
-							<goal>enforce-versions</goal>
-							<goal>set-properties</goal>
-							<goal>retarget-deploy</goal>
-							<goal>update-stage-dependencies</goal>
-						</goals>
-					</execution>
-					<execution>
-						<id>promote-master</id>
-						<goals>
-							<goal>promote-master</goal>
-						</goals>
-						<configuration>
-							<masterBranchPattern>(origin/)?master</masterBranchPattern>
-						</configuration>
-					</execution>
-				</executions>
-			</plugin>
-		</plugins>
-	</build>
-
-</project>
